+++ src/Core/Credentials/Context.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace B24io\Loyalty\SDK\Core\Credentials;
+
+enum Context: string
+{
+    case admin = 'admin';
+    case user = 'user';
+}