+++ src/Core/Response/DTO/ResponseData.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace B24io\Loyalty\SDK\Core\Response\DTO;
+
+readonly class ResponseData
+{
+    public function __construct(
+        /**
+         * @var array<string, mixed>
+         */
+        public array      $result,
+        public Metadata   $metadata,
+        public Pagination $pagination)
+    {
+    }
+}